requirements.txt
numpy>=1.19.5
opencv-python>=4.5.1.48
openslide-python>=1.1.2
matplotlib>=3.3.4
sklearn>=0.24.1
probreg>=0.3.1
pillow>=8.1.0
